Added another explicit conversion for good measure

+++ cfr_br.cpp
@@ -122,7 +122,7 @@ int main(int argc, char* argv[]) {
                 }
                 else {
                     for (int action : actions) {
-                        prob_dist[action] = 1.0 / actions.size();
+                        prob_dist[action] = 1.0 / double(actions.size());
                     }
                 }
             }
@@ -135,7 +135,7 @@ int main(int argc, char* argv[]) {
                 }
                 else {
                     for (int action : actions) {
-                        prob_dist[action] = 1.0 / actions.size();
+                        prob_dist[action] = 1.0 / double(actions.size());
                     }
                 }
             }
